Post-generation moved into its own component
