+++ packages/ui/src/post-maker/index.ts
@@ -5,3 +5,4 @@ export * from './HashtagsSummary'
 export * from './TimeLeft'
 export * from './PostMakerIcon'
 export * from './HashtagProvider'
+export * from './TweetGenAI'
